LPS-57010 Put all the related logic within the same component
