+++ portal-impl/test/integration/com/liferay/portal/security/auth/AuthTokenWhitelistTest.java
@@ -19,11 +19,11 @@ import com.liferay.portal.security.auth.bundle.authtokenignoreactions.TestAuthTo
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.MainServletTestRule;
 import com.liferay.portal.test.rule.SyntheticBundleRule;
+import com.liferay.portal.util.PropsValues;
 
 import java.util.Set;
 
 import org.junit.Assert;
-import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
@@ -31,7 +31,7 @@ import org.junit.Test;
 /**
  * @author Cristina González
  */
-public class AuthTokenWhitelistUtilTest {
+public class AuthTokenWhitelistTest {
 
 	@ClassRule
 	@Rule
@@ -40,9 +40,18 @@ public class AuthTokenWhitelistUtilTest {
 			new LiferayIntegrationTestRule(), MainServletTestRule.INSTANCE,
 			new SyntheticBundleRule("bundle.authtokenignoreactions"));
 
-	@Before
-	public void setUp() {
-		AuthTokenIgnoreActionsRegistry.register(new String[] {"a", "b"});
+	@Test
+	public void testGetPortletCSRFWhitelistActionsDefinedInPortalProperties() {
+		Set<String> portletCSRFWhitelistActions =
+			AuthTokenWhitelistUtil.getPortletCSRFWhitelistActions();
+
+		for (String authTokenIgnoreAction :
+				PropsValues.AUTH_TOKEN_IGNORE_ACTIONS) {
+
+			Assert.assertTrue(
+				"The URL a can't be found in " + portletCSRFWhitelistActions,
+				portletCSRFWhitelistActions.contains(authTokenIgnoreAction));
+		}
 	}
 
 	@Test
@@ -53,27 +62,9 @@ public class AuthTokenWhitelistUtilTest {
 		Assert.assertTrue(
 			"The URL " +
 				TestAuthTokenIgnoreActions.TEST_AUTH_TOKEN_IGNORE_ACTION_URL +
-				" can't be found in the list portletCSRFWhitelistActions",
+				" can't be found in " + portletCSRFWhitelistActions,
 			portletCSRFWhitelistActions.contains(
 				TestAuthTokenIgnoreActions.TEST_AUTH_TOKEN_IGNORE_ACTION_URL));
-
-		Assert.assertTrue(
-			"The URL a can't be found in the list portletCSRFWhitelistActions",
-			portletCSRFWhitelistActions.contains("a"));
-	}
-
-	@Test
-	public void testGetPortletCSRFWhitelistActionsWithRegistry() {
-		Set<String> portletCSRFWhitelistActions =
-			AuthTokenWhitelistUtil.getPortletCSRFWhitelistActions();
-
-		Assert.assertTrue(
-			"The URL a can't be found in the list portletCSRFWhitelistActions",
-			portletCSRFWhitelistActions.contains("a"));
-
-		Assert.assertTrue(
-			"The URL a can't be found in the list portletCSRFWhitelistActions",
-			portletCSRFWhitelistActions.contains("b"));
 	}
 
 }
